(Attempt) to fix some incorrect carry/half-carry implementations
- Note: Some flags (C/H) may be being set incorrectly now....

+++ include/bit.h
@@ -21,9 +21,9 @@ class Bit
 		static void Reset(BYTE &val, BYTE bit);
 		static BYTE Get(BYTE val, BYTE bit);
 		static BYTE DidHalfCarry(BYTE val, BYTE val2, BYTE mask);
-		static WORD DidHalfCarry(WORD val, WORD val2, WORD mask);
-		static BYTE DidCarry(BYTE val, BYTE mask);
-		static WORD DidCarry(WORD val, WORD mask);
+		static BYTE DidHalfCarry(WORD val, WORD val2, WORD mask);
+		static BYTE DidCarry(WORD val, WORD mask);
+		static BYTE DidCarry(int val, WORD mask);
 };
 
 #endif
